Move initialize_test_organization logic to parsec.test_utils
